adding postcard

+++ src/actions/user.action.ts
@@ -54,7 +54,7 @@ export async function getUserByClerkId(clerkId: string) {
 
 export async function getDbUserId() {
   const { userId: clerkId } = await auth();
-  if (!clerkId) throw new Error("Unauthorized");
+  if (!clerkId) return null;
   const user = await getUserByClerkId(clerkId);
   if (!user) throw new Error("User not found");
   return user.id;
@@ -62,6 +62,7 @@ export async function getDbUserId() {
 export async function getRandomUsers() {
   try {
     const userId = await getDbUserId();
+    if (!userId) return [];
     // const userId = "1"
     // get 3 random user excclude ourselves &user that we already follow
     const randomUsers = await prisma.user.findMany({
@@ -101,26 +102,28 @@ export async function getRandomUsers() {
 export async function toggleFollow(targetUserId: string) {
   try {
     const userId = await getDbUserId();
+    if (!userId) return;
+
     if (userId === targetUserId) throw new Error("Cannot follow yourself");
     const existingFollow = await prisma.follows.findUnique({
       where: {
-        followerId_followingId:{
+        followerId_followingId: {
           followerId: userId,
-          followingId: targetUserId
-        }
+          followingId: targetUserId,
+        },
       },
     });
-    if(existingFollow){
+    if (existingFollow) {
       // unfollow
       await prisma.follows.delete({
         where: {
-          followerId_followingId:{
+          followerId_followingId: {
             followerId: userId,
-            followingId: targetUserId
-          }
+            followingId: targetUserId,
+          },
         },
-      })
-    } else{
+      });
+    } else {
       // follow
       await prisma.$transaction([
         prisma.follows.create({
@@ -139,10 +142,10 @@ export async function toggleFollow(targetUserId: string) {
         }),
       ]);
     }
-    revalidatePath("/")
-    return {success: true}
+    revalidatePath("/");
+    return { success: true };
   } catch (error) {
     console.log("Error in toggleFollow", error);
-    return {success: false, error: "Error following user"};
+    return { success: false, error: "Error following user" };
   }
 }
